Revert "Merge tag 'v1.32.0' into smtp-v1.31.2"
This reverts commit d94f533ede7e257437800133e58c2b46bdff0fb0, reversing
changes made to 1801a3d0455a06930ef256c4bc9d9f4fd10ff7ea.

+++ contrib/language/filters/http/source/config.cc
@@ -51,7 +51,7 @@ Http::FilterFactoryCb LanguageFilterFactory::createFilterFactoryFromProtoTyped(
   if (U_FAILURE(errorCode)) {
     throw EnvoyException(fmt::format("Failed to initialize icu::LocaleMatcher::Builder: ICU error "
                                      "code icu::LocaleMatcher::Builder build: {}",
-                                     static_cast<int>(errorCode)));
+                                     errorCode));
   }
 
   auto config = std::make_shared<LanguageFilterConfigImpl>(
