+++ contrib/language/filters/http/source/language_filter.cc
@@ -123,12 +123,10 @@ std::string LanguageFilter::matchValue(icu::Locale& locale) {
     if (U_SUCCESS(errorCode)) {
       return language_tag;
     } else {
-      ENVOY_LOG(error, "ICU error code icu::LocaleMatcher toLanguageTag: {}",
-                static_cast<int>(errorCode));
+      ENVOY_LOG(error, "ICU error code icu::LocaleMatcher toLanguageTag: {}", errorCode);
     }
   } else {
-    ENVOY_LOG(error, "ICU error code icu::LocaleMatcher getBestMatch: {}",
-              static_cast<int>(errorCode));
+    ENVOY_LOG(error, "ICU error code icu::LocaleMatcher getBestMatch: {}", errorCode);
   }
 
   return EMPTY_STRING;
